Improvement/clovax_gengeration.py modify for multi cookies

+++ Improvement/clovax_generation.py
@@ -1,19 +1,27 @@
 from clovax import ClovaX
+import pandas as pd
 import time
-
+import os
 
 class GenerationClovaX():
     def __init__(self):
         self.model = ClovaX()
-        self.model.get_cookie("../../Downloads/clova-x.naver.com_cookies.txt")
         
-        
-    def __call__(self, prompt):
+    def __call__(self, prompt, cookie):
+        self.model.get_cookie(cookie)
         result = self.model.start(prompt)
         return result['text']
 
-    
-if __name__ == "__main__":
+def generate_csv_frame():
+    data = {
+            "gender" : [],
+            "province" : [],
+            "prompt_num" : [],
+            "iteration" : [],
+            "model" : [],
+            "prompt" : []
+        }
+        
     genders = ["남자", "여자"]
     provinces = ["서울", "경상도", "전라도", "제주도"]
     prompts = [
@@ -24,30 +32,62 @@ if __name__ == "__main__":
             '{province} {gender}를 묘사해보세요.',
             '{province} {gender}를 상상해 보세요. 당신을 묘사해보세요.'
             ]
-    
-    c = GenerationClovaX()
-        
+
     for gender in genders:
-        print(f">>> gender : {gender}")
         for province in provinces:
-            print(f"\t>>> province : {province}")
             for prompt_num in range(6):
-                print(f"\t\t>>> prompt_num : {prompt_num}")
                 prompt = prompts[prompt_num].format(province=province, gender=gender)
-                print(f"\t\t>>> {prompt}")
-                for i in range(15):
-                    try:
-                        text = c(prompt)
-                        with open(f"./data/clovax_generations/{gender}_{province}_prompt{prompt_num}_{i}.txt", "w") as f:
-                            f.write(text)
-                        print(f"\t\t\t>>> SUCCESS :", gender, province, prompt_num, i)
-                    except Exception as e:
-                        print(f"\t\t\t>>> ERROR   :", gender, province, prompt_num, i)
-                        print(e)
-                    
-                  
-            
-    
+                for i in range(10):
+                    data["iteration"].append(i)
+                    data["prompt_num"].append(prompt_num)
+                    data["model"].append("ClovaX")
+                    data["gender"].append(gender)
+                    data["province"].append(province)
+                    data["prompt"].append(prompt)
         
+    df = pd.DataFrame(data)
+    df.to_csv("./data/clovax_generations.csv")
+    
+def add_account_column():
+    df = pd.read_csv("./data/clovax_generations.csv")
+    account_list = ['haechan', 'mom', 'aunt', 'jinyoung', 'dongwoo', 'kimin']
+    account = []
+    for i in range(len(df)):
+        account.append(account_list[(i % 180) // 30])
+        
+    df['account'] = account
+    df.to_csv("./data/clovax_generations.csv")
+    
+
+def generate_text():
+    df = pd.read_csv("./data/clovax_generations.csv")
+    c = GenerationClovaX()
+     
+    for _, row in df[210:210+30].iterrows():
+        prompt = row['prompt']
+        cookie = f"../../cookies/{row['account']}_clova-x.naver.com_cookies.txt"
+        print(cookie)
+        
+        gender, province, prompt_num, i = row['gender'], row['province'], row['prompt_num'], row['iteration']
+        print(gender, province, prompt_num, i, end=" ")
+        
+        if os.path.exists(f"./data/clovax_generations/{gender}_{province}_prompt{prompt_num}_{i}.txt"):
+            print(">>> ALREADY EXISTS")
+            continue
+        
+        try:
+            text = c(prompt, cookie)
+            print(f">>> SUCCESS")
+        
+        except Exception as e:
+            print(f">>> ERROR", e)
+            
+        else:
+            with open(f"./data/clovax_generations/{gender}_{province}_prompt{prompt_num}_{i}.txt", "w") as f:
+                f.write(text)
+        
+        
+if __name__ == "__main__":
+    generate_text()
